Update and add frontend unit tests
