+++ frontend/package.json
@@ -44,5 +44,12 @@
   },
   "devDependencies": {
     "eslint-plugin-react": "^7.33.1"
+  },
+  "jest": {
+    "moduleNameMapper": {
+      "swiper/react": "swiper/swiper-react.d.ts",
+      "swiper/modules": "swiper/types/modules/index.d.ts",
+      "swiper/css": "swiper/swiper.min.css"
+    }
   }
 }
